+++ src/main/kotlin/org/jetbrains/intellij/platform/gradleplugin/plugins/IntelliJPlatformBasePlugin.kt
@@ -4,10 +4,19 @@ package org.jetbrains.intellij.platform.gradleplugin.plugins
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.artifacts.Configuration
 import org.gradle.api.model.ObjectFactory
+import org.gradle.api.plugins.JavaPlugin
+import org.gradle.api.plugins.JavaPlugin.COMPILE_ONLY_CONFIGURATION_NAME
+import org.gradle.api.plugins.JavaPlugin.TEST_COMPILE_ONLY_CONFIGURATION_NAME
 import org.gradle.api.provider.ProviderFactory
+import org.gradle.kotlin.dsl.apply
 import org.gradle.kotlin.dsl.create
 import org.jetbrains.intellij.platform.gradleplugin.IntelliJPluginConstants.EXTENSION_NAME
+import org.jetbrains.intellij.platform.gradleplugin.IntelliJPluginConstants.INTELLIJ_PLATFORM_CONFIGURATION_NAME
+import org.jetbrains.intellij.platform.gradleplugin.IntelliJPluginConstants.INTELLIJ_PLATFORM_SOURCES_CONFIGURATION_NAME
+import org.jetbrains.intellij.platform.gradleplugin.IntelliJPluginConstants.JAVA_TEST_FIXTURES_PLUGIN_ID
+import org.jetbrains.intellij.platform.gradleplugin.IntelliJPluginConstants.TEST_FIXTURES_COMPILE_ONLY_CONFIGURATION_NAME
 import org.jetbrains.intellij.platform.gradleplugin.artifacts.transform.applyIntellijPlatformCollectorTransformer
 import org.jetbrains.intellij.platform.gradleplugin.artifacts.transform.applyIntellijPlatformExtractTransformer
 import org.jetbrains.intellij.platform.gradleplugin.checkGradleVersion
@@ -31,10 +40,44 @@ abstract class IntelliJPlatformBasePlugin @Inject constructor(
         checkGradleVersion()
 
         project.repositories.applyIntelliJPlatformSettings(objects, providers)
-        project.applyIntellijPlatformExtractTransformer()
-        project.applyIntellijPlatformCollectorTransformer()
+
+        project.applyPlugins()
+        project.applyConfigurations()
         project.applyExtension()
         project.applyTasks()
+        project.applyTransformers()
+    }
+
+    private fun Project.applyPlugins() {
+        project.plugins.apply(JavaPlugin::class)
+    }
+
+    private fun Project.applyConfigurations() {
+        val intellijPlatformConfiguration = project.configurations.maybeCreate(INTELLIJ_PLATFORM_CONFIGURATION_NAME)
+            .apply {
+                isVisible = false
+                isCanBeConsumed = false
+                isCanBeResolved = true
+                description = "IntelliJ Platform dependency"
+            }
+
+        project.configurations.maybeCreate(INTELLIJ_PLATFORM_SOURCES_CONFIGURATION_NAME)
+            .apply {
+                isVisible = false
+                isCanBeConsumed = false
+                isCanBeResolved = true
+                description = "IntelliJ Platform Sources to be attached to the IntelliJ Platform dependency"
+            }
+
+        fun Configuration.extend() = extendsFrom(intellijPlatformConfiguration)
+
+        with(project.configurations) {
+            getByName(COMPILE_ONLY_CONFIGURATION_NAME).extend()
+            getByName(TEST_COMPILE_ONLY_CONFIGURATION_NAME).extend()
+            project.pluginManager.withPlugin(JAVA_TEST_FIXTURES_PLUGIN_ID) {
+                getByName(TEST_FIXTURES_COMPILE_ONLY_CONFIGURATION_NAME).extend()
+            }
+        }
     }
 
     private fun Project.applyExtension() {
@@ -43,5 +86,10 @@ abstract class IntelliJPlatformBasePlugin @Inject constructor(
 
     private fun Project.applyTasks() {
 
+    }
+
+    private fun Project.applyTransformers() {
+        applyIntellijPlatformExtractTransformer()
+        applyIntellijPlatformCollectorTransformer()
     }
 }